[refactor] BoardQueryService 추상화 추가
