+++ core-service/build.gradle
@@ -51,7 +51,9 @@ dependencies {
     // https://mvnrepository.com/artifact/io.jsonwebtoken/jjwt-jackson
     runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
 
-}
+    // feign
+    implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
+} 
 
 dependencyManagement {
     imports {
