+++ core-service/src/main/java/com/microsoftwo/clother/Be142ndMicrosoftwoClotherApplication.java
@@ -3,9 +3,11 @@ package com.microsoftwo.clother;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
 @SpringBootApplication(scanBasePackages = "com.microsoftwo.clother")
 @EnableDiscoveryClient
+@EnableFeignClients
 public class Be142ndMicrosoftwoClotherApplication {
 
     public static void main(String[] args) {
